requirements.txt
pip==23.2.1
altair==4.2.2
appdirs==1.4.4
attrs==23.1.0
audioread==3.0.0
blinker==1.6.2
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.7
colorama==0.4.6
cycler==0.11.0
decorator==5.1.1
distlib==0.3.6
entrypoints==0.4
filelock==3.11.0
gitdb==4.0.10
GitPython==3.1.31
idna==3.4
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
kiwisolver==1.4.4
lazy_loader==0.2
librosa==0.10.0.post2
markdown-it-py==2.2.0
MarkupSafe==2.1.3
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.0.5
opencv-python==4.7.0.72
packaging==23.1
pafy==0.5.5
Pillow==9.5.0
platformdirs==3.2.0
pooch==1.6.0
protobuf==3.20.3
psutil==5.9.5
pyarrow==11.0.0
pycparser==2.21
pydeck==0.8.1b0
Pygments==2.15.1
Pympler==1.0.1
pyparsing==3.1.0b1
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
requests==2.29.0
rich==13.3.4
seaborn==0.12.2
sentry-sdk==1.23.0
six==1.16.0
smmap==5.0.0
soundfile==0.12.1
soxr==0.3.5
streamlit==1.21.0
thop==0.1.1.post2209072238
threadpoolctl==3.1.0
toml==0.10.2
toolz==0.12.0
tqdm==4.65.0
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
ultralytics==8.0.103
urllib3==1.26.15
validators==0.20.0
virtualenv==20.21.0
watchdog==3.0.0
zipp==3.15.0
Flask==2.0.1
tensorboard==2.6.0
tensorboard-data-server==0.7.1
tensorflow==2.11.0
pymongo==4.5.0
Flask-PyMongo==2.3.0
Flask-Cors==4.0.0